Include Activity Indicator when logging in
When a user is logging in, an activity indicator will show, showing
that the application is waiting for a response from the server.

This deals with #5, although it has been difficult to test on the iOS
simulator due to the good quality of connection (this gives little time
for the indicator to show).