Allow display of a server error at login and in main app.
Previously, if the server was down or a connection error, the app
would crash outright when one tried to launch the app. A similar
situation would occur when server would go down during app usage.

The change in socketOnRecievedData and the sawServerError flag both
determine when the server error alert should be shown.